requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Django==4.0.3
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-js-asset==2.0.0
Pillow==9.0.1
soupsieve==2.3.1
sqlparse==0.4.2
Unidecode==1.3.4
